+++ packages/web-platform/web-tests/server.js
@@ -1,38 +1,23 @@
-import { createLynxView } from '@lynx-js/web-core-server';
+import { executeTemplate } from '@lynx-js/web-core/server';
 import { readFile } from 'node:fs/promises';
 import { fileURLToPath } from 'node:url';
 import path from 'node:path';
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 export async function loadTemplate(caseName) {
-  return JSON.parse(
-    await readFile(
-      path.join(__dirname, 'dist', 'ssr', caseName, 'index.web.json'),
-      'utf-8',
-    ),
+  return await readFile(
+    path.join(__dirname, 'dist', 'ssr', caseName, 'index.web.json'),
   );
 }
 export async function SSR(rawTemplate, caseName, projectName = 'fp-only') {
-  const lynxView = await createLynxView({
-    browserConfig: {
-      pixelRatio: 1,
-      pixelHeight: 800,
-      pixelWidth: 375,
-    },
-    tagMap: {},
-    initData: { mockData: 'mockData' },
-    globalProps: {},
-    template: rawTemplate,
-    templateName: caseName,
-    hydrateUrl: `/dist/ssr/${caseName}/index.web.json`,
-    injectStyles:
-      `@import url("/${projectName}.css"); .injected-style-rules{background:green}`,
-    autoSize: true,
-    lynxViewStyle: 'width:100vw; max-width: 500px;',
-    threadStrategy: 'all-on-ui',
-  });
-  const ssrHtml = await lynxView.renderToString();
-  return ssrHtml.toString('utf-16le');
+  const ssrHtml = executeTemplate(
+    rawTemplate,
+    { mockData: 'mockData' },
+    {},
+    [],
+    'height="auto"',
+  );
+  return ssrHtml;
 }
 export async function genTemplate(caseName, projectName = 'fp-only') {
   const ssrHtml = await SSR(