+++ packages/web-platform/web-tests/tests/react.spec.ts
@@ -95,13 +95,13 @@ test.describe('reactlynx3 tests', () => {
         // @ts-expect-error
         globalThis.lynxView.reload();
       });
-      await wait(100);
+      await wait(1000);
       expect(
         await page.evaluate(() =>
           Array.from(
             document.querySelector('lynx-view')?.shadowRoot?.children || [],
           )
-            .filter(i => i.getAttribute('lynx-tag') === 'page').length
+            .filter(i => i.getAttribute('part') === 'page').length
         ),
       ).toBe(1);
     });
@@ -510,6 +510,23 @@ test.describe('reactlynx3 tests', () => {
         node.scrollTop
       );
       expect(scrollTopAfter).toBeGreaterThan(100);
+    });
+
+    test('basic-main-query-selector-all', async ({ page }, { title }) => {
+      await goto(page, title);
+      await wait(100);
+      const scrollViews = page.locator('scroll-view');
+      const thirdScrollView = scrollViews.nth(2);
+      const scrollTopBefore = await thirdScrollView.evaluate((node) =>
+        node.scrollTop
+      );
+      expect(scrollTopBefore).toBe(0);
+      await page.locator('#tap-me').click();
+      await wait(3000);
+      const scrollTopAfter = await thirdScrollView.evaluate((node) =>
+        node.scrollTop
+      );
+      expect(scrollTopAfter).toBeGreaterThan(50);
     });
 
     // lazy component
@@ -994,13 +1011,6 @@ test.describe('reactlynx3 tests', () => {
     });
   });
   test.describe('apis', () => {
-    test('api-custom-template-loader', async ({ page }, { title }) => {
-      test.skip(isSSR, 'No need to test on SSR');
-      await goto(page, title);
-      await wait(100);
-      const target = page.locator('#target');
-      await expect(target).toHaveCSS('background-color', 'rgb(0, 128, 0)'); // green
-    });
     test('api-animation-event', async ({ page }, { title }) => {
       await goto(page, title);
       await page.locator('#tap1').click();
@@ -1296,7 +1306,7 @@ test.describe('reactlynx3 tests', () => {
       });
       await wait(50);
       expect(message).toContain('fin');
-      expect(page.workers().length).toStrictEqual(0);
+      expect(page.workers().length).toStrictEqual(1);
     });
 
     test('api-error', async ({ page }, { title }) => {
@@ -1500,7 +1510,7 @@ test.describe('reactlynx3 tests', () => {
         document.body.querySelector('lynx-view')?.remove()
       );
       await wait(100);
-      expect(page.workers().length).toBeLessThanOrEqual(0);
+      expect(page.workers().length).toBeLessThanOrEqual(1);
     });
 
     test.describe('api-exposure', () => {
@@ -2143,7 +2153,7 @@ test.describe('reactlynx3 tests', () => {
           document.querySelector('lynx-view')!.setAttribute('height', '100vh');
           document.querySelector('lynx-view')!.setAttribute(
             'style',
-            'width: 100vw; height: 100vh',
+            'display:flex; width: 100vw; height: 100vh',
           );
         });
         await wait(100);
@@ -3566,6 +3576,7 @@ test.describe('reactlynx3 tests', () => {
       test(
         'basic-element-x-swiper-indicator-dots',
         async ({ page }, { title }) => {
+          test.skip(isSSR);
           await goto(page, title);
           await wait(100);
           await diffScreenShot(page, 'x-swiper', 'indicator-dots', 'index', {
@@ -3577,6 +3588,7 @@ test.describe('reactlynx3 tests', () => {
       test(
         'basic-element-x-swiper-indicator-color',
         async ({ page }, { title }) => {
+          test.skip(isSSR);
           await goto(page, title);
           await wait(100);
           await diffScreenShot(page, 'x-swiper', 'indicator-color', undefined, {
@@ -3829,6 +3841,7 @@ test.describe('reactlynx3 tests', () => {
         });
       });
       test('basic-element-x-swiper-autoplay', async ({ page }, { title }) => {
+        test.skip(isSSR);
         await goto(page, title);
         // default duration: 500, interval: 5000
         await wait(5600);
@@ -3857,6 +3870,7 @@ test.describe('reactlynx3 tests', () => {
             browserName === 'firefox',
             'diffScreenShot cost too long time in firefox',
           );
+          test.skip(isSSR);
           await goto(page, title);
           await wait(5600);
           await diffScreenShot(page, 'x-swiper', 'interval-1', undefined, {
@@ -4679,11 +4693,13 @@ test.describe('reactlynx3 tests', () => {
       test(
         'basic-element-list-remove-action',
         async ({ page }, { title }) => {
+          page.on('console', msg => console.log('[PAGE]', msg.text()));
           test.skip(isSSR, 'not support on SSR');
           await goto(page, title);
 
           // Initial state: loading = true
           // Expected: 1, 2, 3, 5
+          await wait(500);
           await expect(page.locator('list-item').count()).resolves.toBe(4);
           let ids = await page.locator('list-item').evaluateAll((items) =>
             items.map((i) => i.id)