+++ packages/webpack/react-webpack-plugin/test/cases/main-thread/lazy-bundle-sourcemap/lazy.jsx
@@ -1,0 +1,3 @@
+export default function Lazy() {
+  return <view>Lazy Loaded</view>;
+}