+++ packages/webpack/react-webpack-plugin/test/cases/worklet-runtime/chunk/a.jsx
@@ -2,10 +2,16 @@ export function a2() {
   const onTapMT = () => {
     'main thread';
   };
+  const onLongPressMT = () => {
+    'main thread';
+  };
 
   return (
     <view>
-      <text bindtap={onTapMT}>
+      <text
+        bindtap={onTapMT}
+        bindlongpress={onLongPressMT}
+      >
         hello world
       </text>
     </view>